+++ packages/terra-date-picker/src/DatePicker.jsx
@@ -61,6 +61,10 @@ const propTypes = {
    */
   onClickOutside: PropTypes.func,
   /**
+   * A callback function triggered when the date input receives focus. function(event)
+   */
+  onFocus: PropTypes.func,
+  /**
    * A callback function to execute when a date is selected from within the picker.
    */
   onSelect: PropTypes.func,
@@ -91,6 +95,7 @@ const defaultProps = {
   inputAttributes: undefined,
   maxDate: undefined,
   minDate: undefined,
+  onFocus: undefined,
   onChange: undefined,
   onChangeRaw: undefined,
   onClickOutside: undefined,
@@ -193,8 +198,8 @@ class DatePicker extends React.Component {
   }
 
   handleOnInputFocus(event) {
-    if (this.onInputFocus) {
-      this.onInputFocus(event);
+    if (this.props.onFocus) {
+      this.props.onFocus(event);
     }
 
     if (!this.isDefaultDateAcceptable) {
@@ -241,6 +246,7 @@ class DatePicker extends React.Component {
       maxDate,
       minDate,
       name,
+      onFocus,
       onChange,
       onChangeRaw,
       onClickOutside,
@@ -253,10 +259,8 @@ class DatePicker extends React.Component {
     } = this.props;
 
     this.onCalendarButtonClick = customProps.onCalendarButtonClick;
-    this.onInputFocus = customProps.onInputFocus;
 
     delete customProps.onCalendarButtonClick;
-    delete customProps.onInputFocus;
 
     const { intl } = this.context;
     const todayString = intl.formatMessage({ id: 'Terra.datePicker.today' });
@@ -274,11 +278,11 @@ class DatePicker extends React.Component {
         onChange={this.handleChange}
         onChangeRaw={this.handleChangeRaw}
         onClickOutside={this.handleOnClickOutside}
+        onFocus={this.handleOnInputFocus}
         onSelect={this.handleOnSelect}
         required={required}
         customInput={(
           <DateInput
-            onInputFocus={this.handleOnInputFocus}
             onCalendarButtonClick={this.handleOnCalendarButtonClick}
             inputAttributes={inputAttributes}
             releaseFocus={releaseFocus}
@@ -301,6 +305,7 @@ class DatePicker extends React.Component {
         dropdownMode="select"
         showMonthDropdown
         showYearDropdown
+        preventOpenOnFocus
         name={name}
       />
     );
@@ -312,11 +317,11 @@ class DatePicker extends React.Component {
         onChange={this.handleChange}
         onChangeRaw={this.handleChangeRaw}
         onClickOutside={this.handleOnClickOutside}
+        onFocus={this.handleOnInputFocus}
         onSelect={this.handleOnSelect}
         required={required}
         customInput={(
           <DateInput
-            onInputFocus={this.handleOnInputFocus}
             onCalendarButtonClick={this.handleOnCalendarButtonClick}
             inputAttributes={inputAttributes}
             releaseFocus={releaseFocus}
@@ -347,6 +352,7 @@ class DatePicker extends React.Component {
         dropdownMode="select"
         showMonthDropdown
         showYearDropdown
+        preventOpenOnFocus
         name={name}
       />
     );